requirements.txt
streamlit~=0.87.0
pydeck~=0.7.0
pandas~=1.3.2
matplotlib~=3.4.2
numpy~=1.19.5
seaborn~=0.11.2
plotly~=5.3.1
Pillow~=8.2.0